chore: migrate /api/team/create
